Rename the preprocess to postprocess
